updating tests and docs
