+++ docs/configureArquillian.md
@@ -2,8 +2,8 @@
 
 The `configureArquillian` task is used to easily integrate Arquillian and the Arquillian Liberty Managed and Remote containers with your existing Liberty project. It provides two major advantages:
 
-1. The ability to set `arquillian.xml` container properties directly within the `pom.xml`.
-2. Automatic generation of `arquillian.xml` when using the Arquillian Liberty Managed container. 
+1. The ability to set `arquillian.xml` container properties directly within the `build.gradle`.
+2. Automatic generation of `arquillian.xml` when using the Arquillian Liberty Managed container.
 
 For more convenient integration with the Arquillian Liberty Managed/Remote containers and JUnit/TestNG, you can use this task with the [Arquillian Liberty Dependency Bundles](https://github.com/wasdev/arquillian-liberty-dependencies).
 
@@ -11,15 +11,15 @@ For more convenient integration with the Arquillian Liberty Managed/Remote conta
 
 #### Automatic generation of `arquillian.xml`
 
-For both Arquillian Liberty Managed and Remote containers, the `configureArquillian` task can automatically generate and configure the `arquillian.xml` that usually exists in the `src/test/resources` directory, eliminating the need for the user to manually create the file in projects with only a single Arquillian container. 
+For both Arquillian Liberty Managed and Remote containers, the `configureArquillian` task can automatically generate and configure the `arquillian.xml` that usually exists in the `src/test/resources` directory, eliminating the need for the user to manually create the file in projects with only a single Arquillian container.
 
-Furthermore, when using the Arquillian Liberty Managed container, `configureArquillian` will also automatically specify the three required configuration parameters: `wlpHome`, `serverName`, and `httpPort`. These are set to values specified in the `liberty-gradle-plugin`. 
+Furthermore, when using the Arquillian Liberty Managed container, `configureArquillian` will also automatically specify the three required configuration parameters: `wlpHome`, `serverName`, and `httpPort`. These are set to values specified in the `liberty-gradle-plugin`.
 
-The `configureArquillian` task will perform configuration for the Liberty Managed or Remote container based on which container exists on the classpath. Arquillian does not allow for more than one container on the classpath. In the event of neither container existing on the classpath, the `configureArquillian` task will default to providing configuration for the Liberty Managed container. 
+The `configureArquillian` task will perform configuration for the Liberty Managed or Remote container based on which container exists on the classpath. Arquillian does not allow for more than one container on the classpath. In the event of neither container existing on the classpath, the `configureArquillian` task will default to providing configuration for the Liberty Managed container.
 
 ### dependsOn
 
-`configureArquillian` depends on `libertyCreate` and `processTestResources`.
+`configureArquillian` depends on `installApps` and `processTestResources`.
 
 ### Properties
 
@@ -38,10 +38,10 @@ https://docs.jboss.org/author/display/ARQ/WLP+V8.5+-+Remote
 
 ### Examples
 
-This is an example of using `configureArquillian` in a Managed container environment. Specify the following tasks below the `liberty` closure in your `build.gradle`:
+This is an example of using `configureArquillian` in a Managed container environment. Specify the following tasks below the `liberty` closure in your `build.gradle`. The properties used by `configureArquillian` are set inside of the `arquillianConfiguration` extension:
 
 ```
-configureArquillian {
+arquillianConfiguration {
     arquillianProperties = ['outputToConsole': 'true']
     skipIfArquillianXmlExists = true
 }
@@ -56,7 +56,7 @@ test {
 }
 ```
 
-Note that `test` depends on `configureArquillian` to ensure that the `arquillian.xml` will be generated before integration tests are run. 
+Note that `test` depends on `configureArquillian` to ensure that the `arquillian.xml` will be generated before integration tests are run.
 
 You can verify that the `arquillian.xml` is being generated properly by viewing it at `build/resources/test/arquillian.xml`. You'll know that the file was generated by the `configureArquillian` task if the following comment is at the end of the file:
